Add versioning to user avatars
