+++ user/src/main/proto/user.proto
@@ -30,6 +30,7 @@ message UserProfile {
   security.ProofOfWork proofOfWork = 3;
   string terms = 4;
   string statement = 5;
+  string avatarVersion = 6;
 }
 
 message UserIdentity {
